requirements.txt
inflect==0.2.5
oauthlib==1.1.2
python-dateutil==2.5.3
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
tweepy==3.5.0
